requirements.txt
protobuf==3.20.3
PyPDF2==3.0.1
python-dotenv==1.1.1
streamlit==1.12.0
standard-imghdr==3.13.0
altair==4.2.2
google-genai==1.30.0


